refactor: extract import + drop-folder routes into routes/import.ts
Pure structural move out of createApp; no handler logic changed. Moves the
30 evidence-import + drop-folder routes (unified /import, /import-file, every
/import-<kind>, /import-meta, the #76 import undo/redo, and the drop-folder
/drop/run-pending + /drop-status) into a new routes/import.ts.

Graduated onto RouteContext (genuinely shared with createApp code that stays —
the Velociraptor bundle collector, the drop poller, and the push/tool paths):
dispatchImport, demoteForensicForCase, resynthesizeInBackground,
pushImportCheckpoint, applyWhitelistToCase, applyNsrlToCase,
applyDeobfuscationToCase, moveDropFile (stable methods), plus the ImportBase
type. The drop-watcher state (dropSeen / dropScanning / dropPendingLogged) is
SHARED with the scanCaseDrops poller that stays in createApp, so it was
graduated as live accessors (the poller keeps mutating the same Maps/Set) rather
than moved. resolveToolForExt + dropDirOf + restoreImportState + dropPollMs are
reconstructed module-locally; restoreImportState's createApp copy (only the moved
undo/redo used it) was removed. Reused pre-existing ctx members: store, options,
serverLogger, recordImportFailure, recordAiError, hasAiProvider, getControl,
runToolAndIngest, resolveImportKind + liveToolConfigs + customTools +
dropWatchEnabled. Dead per-format parser imports pruned from server.ts.

tsc clean; route inventory unchanged (316); tests/server green (400), including
importerRoutes / importMissingCase / importUndoRoutes / veloImportExternal.
